create pkg/boottest to make JSON-comparison reusable for GRUB tests
